Relocate sgxmeasure.c/.h
