+++ common/sgx/sgxmeasure.h
